requirements.txt
Django==1.10
django-calaccess-raw-data==1.4.0rc9
psycopg2==2.6.1
boto3==1.3.1
-e git+https://github.com/syapse/django-storages.git@boto3-new#egg=django-storages
django-bakery==0.8.9
six
Fabric
flake8
coveralls
configparser
